Move offset log tests to dedicated file
