+++ sql/core/src/test/scala/org/apache/spark/sql/execution/streaming/OffsetSeqLogSuite.scala
@@ -223,81 +223,6 @@ class OffsetSeqLogSuite extends SharedSparkSession {
     }
   }
 
-  test("Protobuf extensions support disabled by default for new queries") {
-    val offsetSeqMetadata =
-      OffsetSeqMetadata(batchWatermarkMs = 0, batchTimestampMs = 0, spark.conf)
-    assert(
-      offsetSeqMetadata.conf.get(SQLConf.PROTOBUF_EXTENSIONS_SUPPORT_ENABLED.key) ===
-        Some(false.toString))
-  }
-
-  test("Protobuf extensions support enabled when session sets it to true") {
-    val protobufExtConf = SQLConf.PROTOBUF_EXTENSIONS_SUPPORT_ENABLED.key
-    withSQLConf(protobufExtConf -> true.toString) {
-      val offsetSeqMetadata =
-        OffsetSeqMetadata(batchWatermarkMs = 0, batchTimestampMs = 0, spark.conf)
-      assert(offsetSeqMetadata.conf.get(protobufExtConf) === Some(true.toString))
-    }
-  }
-
-  test(
-    "Protobuf extensions support uses default false for old checkpoint when enabled in session") {
-    val protobufExtConf = SQLConf.PROTOBUF_EXTENSIONS_SUPPORT_ENABLED.key
-    withSQLConf(protobufExtConf -> true.toString) {
-      val existingChkpt = "offset-log-version-2.1.0"
-      val (_, offsetSeq) = readFromResource(existingChkpt)
-      val offsetSeqMetadata = offsetSeq.metadataOpt.get
-      // Not present in existing checkpoint
-      assert(offsetSeqMetadata.conf.get(protobufExtConf) === None)
-
-      val clonedSqlConf = spark.sessionState.conf.clone()
-      OffsetSeqMetadata.setSessionConf(offsetSeqMetadata, clonedSqlConf)
-      assert(!clonedSqlConf.getConf(SQLConf.PROTOBUF_EXTENSIONS_SUPPORT_ENABLED))
-    }
-  }
-
-  test(
-    "Protobuf extensions support uses default false for old checkpoint when unset in session") {
-    val existingChkpt = "offset-log-version-2.1.0"
-    val (_, offsetSeq) = readFromResource(existingChkpt)
-    val offsetSeqMetadata = offsetSeq.metadataOpt.get
-    val protobufExtConf = SQLConf.PROTOBUF_EXTENSIONS_SUPPORT_ENABLED.key
-    // Not present in existing checkpoint
-    assert(offsetSeqMetadata.conf.get(protobufExtConf) === None)
-
-    val clonedSqlConf = spark.sessionState.conf.clone()
-    OffsetSeqMetadata.setSessionConf(offsetSeqMetadata, clonedSqlConf)
-    assert(!clonedSqlConf.getConf(SQLConf.PROTOBUF_EXTENSIONS_SUPPORT_ENABLED))
-  }
-
-  test("protobuf extensions support in existing checkpoint takes precedence over session value") {
-    val protobufExtConf = SQLConf.PROTOBUF_EXTENSIONS_SUPPORT_ENABLED.key
-
-    // Enabled when checkpoint = enabled and session = disabled
-    withSQLConf(protobufExtConf -> false.toString) {
-      val offsetSeqMetadata = OffsetSeqMetadata(
-        batchWatermarkMs = 0,
-        batchTimestampMs = 0,
-        Map(protobufExtConf -> true.toString))
-
-      val clonedSqlConf = spark.sessionState.conf.clone()
-      OffsetSeqMetadata.setSessionConf(offsetSeqMetadata, clonedSqlConf)
-      assert(clonedSqlConf.getConf(SQLConf.PROTOBUF_EXTENSIONS_SUPPORT_ENABLED))
-    }
-
-    // Disabled when checkpoint = disabled and session = enabled
-    withSQLConf(protobufExtConf -> true.toString) {
-      val offsetSeqMetadata = OffsetSeqMetadata(
-        batchWatermarkMs = 0,
-        batchTimestampMs = 0,
-        Map(protobufExtConf -> false.toString))
-
-      val clonedSqlConf = spark.sessionState.conf.clone()
-      OffsetSeqMetadata.setSessionConf(offsetSeqMetadata, clonedSqlConf)
-      assert(!clonedSqlConf.getConf(SQLConf.PROTOBUF_EXTENSIONS_SUPPORT_ENABLED))
-    }
-  }
-
   test("OffsetMap golden file compatibility test - VERSION_2 format") {
     val (batchId, offsetSeq) = readFromResource("offset-map")
     assert(batchId === 3)
